assignment4/test/requirements.txt
cvxopt
matplotlib
nose2
numpy
Pillow
pycodestyle
scikit-image
scikit-learn
scipy
